chore: split up actions and reducers
