+++ streetfleet/src/config.js
@@ -1,6 +1,6 @@
 const config = {
  googleMapURL: "https://maps.googleapis.com/maps/api/js?key=AIzaSyAGaBSySyhmv8w1TXxgQO9hx6ZgOPWZsZE&v=3.exp&libraries=geometry,drawing,places",
  baseUrl: 'https://private-040f98-streetfleet.apiary-mock.com/api/v1',
- //baseUrl: 'http://localhost:4000/api/v1',
+//  baseUrl: 'http://localhost:4000/api/v1',
 }
 export default config;
