+++ streetfleet/src/store/configureStore.js
@@ -1,21 +1,20 @@
-import { createStore, applyMiddleware } from 'redux';
+import { createStore, applyMiddleware, compose, combineReducers } from 'redux';
 // import { createLogger } from 'redux-logger';
-import reducers from './reducers';
+import reducer from './reducers/index';
+//import reducer from './reducers_old';
+//import * as reducers from './reducers';
 import api from './middlewares/api';
 import logger from './middlewares/logger';
+import config from './../config';
 
-
-
-
+//const reducer = combineReducers(reducers);
+const composeEnhancers = window.__REDUX_DEVTOOLS_EXTENSION_COMPOSE__ || compose;
 const configureStore = () => {
-  // const logger = createLogger();
-
+ // const logger = createLogger();
   const store = createStore(
-    reducers,
-    window.__REDUX_DEVTOOLS_EXTENSION__ && window.__REDUX_DEVTOOLS_EXTENSION__(),
-    applyMiddleware(api, logger),
+    reducer,
+    composeEnhancers(applyMiddleware(api(config.baseUrl),logger)),
   );
-
   return store;
 };
